third_party/WebKit/LayoutTests/crypto/subtle/aes-cbc/unwrap-failures-expected.txt

Test unwrapping an RSA key with AES-CBC.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

Importing an unwrapping key...
PASS unwrappingKey.algorithm.name is 'AES-CBC'
Unwrapping a key...
error is: DataError: The required JWK member "kty" was missing
PASS successfullyParsed is true

TEST COMPLETE
